+++ apps-tools/calib_app/src/filter_logic.cpp
@@ -31,14 +31,14 @@ CFilter_logic::Ptr CFilter_logic::Create(CCalibMan::Ptr _calib_man)
     return std::make_shared<CFilter_logic>(_calib_man);
 }
 
-bool compare(CFilter_logic::GridItem i1, CFilter_logic::GridItem i2) 
-{ 
-    return ((i1.value_raw  * i1.value_raw + i1.deviationFromAVG * i1.deviationFromAVG) < (i2.value_raw  * i2.value_raw + i2.deviationFromAVG * i2.deviationFromAVG)); 
-//    return ((i1.value  * i1.value + i1.deviationFromAVG * i1.deviationFromAVG) < (i2.value  * i2.value + i2.deviationFromAVG * i2.deviationFromAVG)); 
-} 
+bool compare(CFilter_logic::GridItem i1, CFilter_logic::GridItem i2)
+{
+    return ((i1.value_raw  * i1.value_raw + i1.deviationFromAVG * i1.deviationFromAVG) < (i2.value_raw  * i2.value_raw + i2.deviationFromAVG * i2.deviationFromAVG));
+//    return ((i1.value  * i1.value + i1.deviationFromAVG * i1.deviationFromAVG) < (i2.value  * i2.value + i2.deviationFromAVG * i2.deviationFromAVG));
+}
 
 bool IsContain(std::vector<CFilter_logic::GridItem> &_list, CFilter_logic::GridItem i1){
-    for (int i = 0 ; i < _list.size() ; i++){
+    for (size_t i = 0 ; i < _list.size() ; i++){
         if (_list[i].aa == i1.aa && _list[i].bb == i1.bb)
             return true;
     }
@@ -102,13 +102,13 @@ void CFilter_logic::setCalibMode(int _mode){
 
 void CFilter_logic::print(){
     sort();
-    for(int i = 0 ; i < m_grid.size() ; i++){
-        printf("%d {AA: %6x, BB: %6x, V: %2.6f, Vraw: %2.6f, D: %3.6f, index: %d}\n",i,m_grid[i].aa,m_grid[i].bb,m_grid[i].value,m_grid[i].value_raw,m_grid[i].deviationFromAVG,m_grid[i].index);
+    for(size_t i = 0 ; i < m_grid.size() ; i++){
+        printf("%d {AA: %6x, BB: %6x, V: %2.6f, Vraw: %2.6f, D: %3.6f, index: %lld}\n",i,m_grid[i].aa,m_grid[i].bb,m_grid[i].value,m_grid[i].value_raw,m_grid[i].deviationFromAVG,m_grid[i].index);
     }
 }
 
 int CFilter_logic::setCalibParameters(){
-    for(int i = 0 ; i < m_grid.size() ; i++){
+    for(size_t i = 0 ; i < m_grid.size() ; i++){
         if (m_grid[i].calculate == false) {
             m_calib_man->setCalibValue(m_grid[i].ch, F_AA_CH, m_grid[i].aa);
             m_calib_man->setCalibValue(m_grid[i].ch, F_BB_CH, m_grid[i].bb);
@@ -135,21 +135,21 @@ int CFilter_logic::getCalibCount(){
 
 int CFilter_logic::getCalibDone(){
     int x = 0;
-    for(int i = 0 ; i < m_grid.size() ; i++){
+    for(size_t i = 0 ; i < m_grid.size() ; i++){
         if (m_grid[i].calculate) x++;
     }
     return x;
 }
 
 int CFilter_logic::setCalculatedValue(COscilloscope::DataPassAutoFilter item){
-    for(int i = 0 ; i < m_grid.size() ; i++){
-        if (m_grid[i].aa == item.f_aa) 
+    for(size_t i = 0 ; i < m_grid.size() ; i++){
+        if (m_grid[i].aa == item.f_aa)
             if (m_grid[i].bb == item.f_bb)
                 if (m_grid[i].ch == item.cur_channel)
                     if (!m_grid[i].calculate)
                     {
                         m_grid[i].value = item.calib_value;
-                        m_grid[i].value_raw = item.calib_value_raw;                        
+                        m_grid[i].value_raw = item.calib_value_raw;
                         m_grid[i].calculate = true;
                         m_grid[i].deviationFromAVG = item.deviation;
     //                    std::cout  << "Cur step " << i <<  " aa = " << m_grid[i].aa << " bb = " << m_grid[i].bb << " VALUE " << m_grid[i].value <<  std::endl;
@@ -162,19 +162,19 @@ int CFilter_logic::setCalculatedValue(COscilloscope::DataPassAutoFilter item){
 void CFilter_logic::removeHalfCalib(){
     sort();
     m_grid.erase (m_grid.end() - m_grid.size() / 3.0 , m_grid.end());
-    if (m_grid.size() > 0) 
+    if (m_grid.size() > 0)
         m_lastGood = m_grid[0];
 }
 
 CFilter_logic::GridItem GetNewItem(std::vector<CFilter_logic::GridItem> _list,int _index){
     std::vector<CFilter_logic::GridItem> new_grid;
-    for(int i = 0 ; i < _list.size() ; i++) {
+    for(size_t i = 0 ; i < _list.size() ; i++) {
         if (_list[i].index == _index){
             new_grid.push_back(_list[i]);
         }
     }
     if (new_grid.size() > 0) {
-        std::sort(new_grid.begin(), new_grid.end(), compare); 
+        std::sort(new_grid.begin(), new_grid.end(), compare);
         return new_grid[0];
     }
     CFilter_logic::GridItem item;
@@ -182,7 +182,7 @@ CFilter_logic::GridItem GetNewItem(std::vector<CFilter_logic::GridItem> _list,in
     return item;
 }
 
-std::vector<CFilter_logic::GridItem> Split(CFilter_logic::GridItem _item,double _step){   
+std::vector<CFilter_logic::GridItem> Split(CFilter_logic::GridItem _item,double _step){
     std::vector<CFilter_logic::GridItem> new_grid;
     if (_step == 0) return new_grid;
     for(double i = 1.0 - (_step / 100.0) ; i <= 1 + (_step / 100.0) ; i += (_step / 100.0))
@@ -204,13 +204,13 @@ int CFilter_logic::nextSetupCalibParameters(){
    if (m_percent < PERCENT_MIN) return -1;
    std::vector<GridItem> new_grid;
    std::vector<int64_t>  index_list;
-   for(int i = 0 ; i < m_grid.size() ; i++){
+   for(size_t i = 0 ; i < m_grid.size() ; i++){
        if (std::find(index_list.begin(), index_list.end(), m_grid[i].index) == index_list.end()){
            index_list.push_back(m_grid[i].index);
        }
    }
-   m_percent /= STEPS_BB; 
-   for(int i = 0 ; i < index_list.size() ; i++){
+   m_percent /= STEPS_BB;
+   for(size_t i = 0 ; i < index_list.size() ; i++){
        auto item = GetNewItem(m_grid,index_list[i]);
        auto split_grid = Split(item,m_percent);
        new_grid.insert(new_grid.end(), split_grid.begin(), split_grid.end());
@@ -221,7 +221,7 @@ int CFilter_logic::nextSetupCalibParameters(){
 }
 
 void CFilter_logic::sort(){
-    std::sort(m_grid.begin(), m_grid.end(), compare); 
+    std::sort(m_grid.begin(), m_grid.end(), compare);
 }
 
 int CFilter_logic::calcProgress(){
@@ -230,7 +230,7 @@ int CFilter_logic::calcProgress(){
     double max_range = 0;
     while(p >= PERCENT_MIN){
         max_range++;
-        p /= STEPS_BB; 
+        p /= STEPS_BB;
         if (p >= m_percent) cur++;
     }
     auto progress = cur / max_range * 100.0;
@@ -241,7 +241,7 @@ int CFilter_logic::calcProgress(){
 
 int CFilter_logic::calibPP(COscilloscope::DataPassAutoFilter item,float _nominal){
     if (item.is_valid == false) return 0;
-    if (m_calibAmpl == 0) return -2;     
+    if (m_calibAmpl == 0) return -2;
     if (m_oldcalibAmpl != -1){
         if (m_oldPP != item.f_pp) return 0;
         float dir = 1;
@@ -262,7 +262,7 @@ int CFilter_logic::calibPP(COscilloscope::DataPassAutoFilter item,float _nominal
         m_calib_man->setCalibValue(item.cur_channel,F_KK_CH,item.f_kk);
         m_calib_man->updateCalib();
         m_calib_man->updateAcqFilter(item.cur_channel);
-        if (m_calibAmpl == 0) return -2; 
+        if (m_calibAmpl == 0) return -2;
     }
     m_oldcalibAmpl = item.ampl;
     m_oldPP = item.f_pp;
